report-build-system/requirements.txt
# Core (always needed)
PyYAML>=6.0

# python-docx engine (optional — only needed when docx-engine: python-docx)
python-docx>=1.1.0
lxml>=4.9.0
